Feature push (early days)
Still working out the best way to handle the library - really worried it's quite forceful in it's method.
Everything returns as generic allowing easy access by array of by object notation.
Collections are returned when an entire array is passed in though.
Errors are now thrown when something goes wrong